speechrecognition/requirements.txt
nltk
numpy
coverage
pytest
jupyter
pandas
pylint
matplotlib
matrixprofile-ts